+++ src/history/test/SerializeTests.cpp
@@ -12,41 +12,30 @@ using namespace stellar;
 
 TEST_CASE("Serialization round trip", "[history]")
 {
-    std::vector<std::string> inputFiles = {
+    std::vector<std::string> testFiles = {
         "stellar-history.testnet.6714239.json",
         "stellar-history.livenet.15686975.json",
         "stellar-history.testnet.6714239.networkPassphrase.json"};
-    std::vector<std::string> expectedOutputFiles = {
-        "stellar-history.testnet.6714239.json",
-        "stellar-history.livenet.15686975.json",
-        "expected-output-stellar-history.testnet.6714239.networkPassphrase.json"};
-    for (int i = 0; i < inputFiles.size(); i++)
+    for (int i = 0; i < testFiles.size(); i++)
     {
         std::string fnPath = "testdata/";
-        std::string inputPath = fnPath + inputFiles[i];
-        std::string expectedOutputPath = fnPath + expectedOutputFiles[i];
-        SECTION("Serialize " + inputPath)
+        std::string testFilePath = fnPath + testFiles[i];
+        SECTION("Serialize " + testFilePath)
         {
-            std::ifstream in(inputPath);
+            std::ifstream in(testFilePath);
             in.exceptions(std::ios::badbit);
-            std::string input((std::istreambuf_iterator<char>(in)),
-                              std::istreambuf_iterator<char>());
-
-            std::ifstream expectedOutputStream(expectedOutputPath);
-            expectedOutputStream.exceptions(std::ios::badbit);
-            std::string expectedOutput(
-                (std::istreambuf_iterator<char>(expectedOutputStream)),
-                std::istreambuf_iterator<char>());
+            std::string hasString((std::istreambuf_iterator<char>(in)),
+                                  std::istreambuf_iterator<char>());
 
             // Test fromString
             HistoryArchiveState has;
-            has.fromString(input);
-            REQUIRE(expectedOutput == has.toString());
+            has.fromString(hasString);
+            REQUIRE(hasString == has.toString());
 
             // Test load
             HistoryArchiveState hasLoad;
-            hasLoad.load(inputPath);
-            REQUIRE(expectedOutput == hasLoad.toString());
+            hasLoad.load(testFilePath);
+            REQUIRE(hasString == hasLoad.toString());
         }
     }
 }